refactor group service and controller

+++ app/Services/GroupService.php
@@ -24,7 +24,7 @@ class GroupService extends Service
             'user_id' => $adminId,
         ]);
         
-        return $group;
+        return ['message' => 'Group created successfully', 'group' => $group];
     }
 
     public function getGroupById($groupId) 
@@ -44,14 +44,45 @@ class GroupService extends Service
         return $group->files()->where('status', '=', 0)->exists();
     }
 
-    public function deleteGroup($group)
+    public function deleteGroup($authenticatedUserId, $groupId)
     {
-        return $group->delete();
+        $group = $this->getGroupById($groupId);
+        if (!$group) {// Check if the group exists
+            return ['message' => 'Group not found', 'status' => 404];
+        }
+        // Check if the authenticated user owns the group
+        if ($authenticatedUserId !== $group->admin_id) {
+            return ['message' => 'You are not an admin of this group', 'status' => 401];
+        }
+
+        $filesWithStatusOne = $this->isGroupHasBookedFile($group);
+        if ($filesWithStatusOne) {
+            return ['message' => 'Cannot delete group with associated files', 'status' => 422];
+        }
+
+        $group->delete();
+        return ['message' => 'Group , associated members and the files in this group is deleted successfully'];
     }
 
-    public function addMember($group, $userToAddId)
+    public function addMember($authenticatedUserId, $groupId, $userToAdd)
     { 
-        return $group->group_member()->create(['user_id' => $userToAddId]);
+        $group = $this->getGroupById($groupId);
+        $user = $this->userService->getUserByEmailOrUsername($userToAdd);
+        if (!$group || !$user) { // Check if the group and user exist
+            return ['message' => 'Group or user not found', 'status' => 404];
+        }
+
+        // Check if the authenticated user is the group admin who created the group
+        if ($authenticatedUserId !== $group->admin_id) {
+            return ['message' => 'Unauthorized. You are not the group admin who created the group.', 'status' => 401];
+        }
+
+        $isMemberExist = $this->isMemberExist($group->id, $user->id);
+        if ($isMemberExist) {//check if the user want to add is already exist in the group
+            return ['message' => 'this user is already exist in this group', 'status' => 400];
+        }
+        $member = $group->group_member()->create(['user_id' => $user->id]);
+        return ['message' => 'Group member added successfully', 'member' => $member];
     }
     
     public function getUserGroups($userId)
@@ -64,7 +95,6 @@ class GroupService extends Service
         $formattedGroups = $groups->map(function ($group) {
             $num=Group_member::where('group_id','=',$group->id)->count();
             return [
-                'msg'=>'hi',
                 'group_id' => $group->id,
                 'name' => $group->name,
                 'admin_id' => $group->admin_id,
@@ -72,7 +102,7 @@ class GroupService extends Service
             ];
         });
 
-        return $formattedGroups;
+        return ['message' => 'All groups', 'allGroups' => $formattedGroups];
     }
 
 
@@ -87,8 +117,18 @@ class GroupService extends Service
         ];
     }
 
-    public function getGroupMembers($groupId)
+    public function getGroupMembers($authenticatedUserId, $groupId)
     {
+        $group = $this->getGroupById($groupId);
+        if (!$group) {// Check if the group exists
+            return ['message' => 'Group not found', 'status' => 404];
+        }
+
+        $isMemberExist = $this->isMemberExist($group->id, $authenticatedUserId);
+        if (!$isMemberExist) {//check if the authenticated user is a member in the group
+            return ['message' => 'you are not a member in this group', 'status' => 400];
+        }
+        $admin = $this->getGroupAdmin($group->admin_id);
         $groupMembers = Group_member::query()->where('group_id', $groupId)
             ->with(['user:id,username,email'])->get(['user_id']);
         $userDetails = $groupMembers->map(function ($groupMember) {
@@ -99,7 +139,7 @@ class GroupService extends Service
             ];
         });
 
-        return $userDetails;
+        return ['message' => 'Users in this group', 'group' => $group, 'userDetails' => $userDetails, 'admin' => $admin];
     }
 
     public function isMemberHasBookedFiles($group_id, $memberId)
@@ -111,11 +151,24 @@ class GroupService extends Service
             ->exists();
     }
 
-    public function deleteMember($groupId, $memberId)
+    public function deleteMember($authenticatedUserId, $groupId, $memberId)
     {
+        $group = $this->getGroupById($groupId);
+        $isMemberExist = $this->isMemberExist($groupId, $memberId);
+        if (!$isMemberExist) {
+            return ['message' => 'Group member not found in the specified group', 'status' => 404];
+        }
+        if ($authenticatedUserId !== $group->admin_id) {
+            return ['message' => 'Unauthorized. You are not the group admin who created the group.', 'status' => 401];
+        }
+        $bookedFilesExist = $this->isMemberHasBookedFiles($groupId, $memberId);
+        if ($bookedFilesExist) {
+            return ['message' => 'Sorry. You cannot delete this member because they have booked a file.', 'status' => 401];
+        }
         $groupMember = Group_member::where('group_id', $groupId)
             ->where('user_id', $memberId)
             ->first();
-        return $groupMember->delete();
+        $groupMember->delete();
+        return ['message' => 'The member deleted successfully'];
     }
 }
